+++ types/test_util.go
@@ -6,7 +6,7 @@ import (
 	"github.com/pkg/errors"
 )
 
-func MakeCommit(blockID BlockID, height int64, round int,
+func MakeCommit(blockID BlockID, height int64, round int32,
 	voteSet *VoteSet, validators []PrivValidator, now time.Time) (*Commit, error) {
 
 	// all sign
@@ -17,7 +17,7 @@ func MakeCommit(blockID BlockID, height int64, round int,
 		}
 		vote := &Vote{
 			ValidatorAddress: pubKey.Address(),
-			ValidatorIndex:   i,
+			ValidatorIndex:   int32(i),
 			Height:           height,
 			Round:            round,
 			Type:             PrecommitType,
